day20 - avoid i128 to speed up

+++ src/day20.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use std::collections::VecDeque;
 use std::time::{Duration, Instant};
 
-fn run_once(deq: &mut VecDeque<(i64, i128)>, input: &[(i64, i128)]) {
+fn run_once(deq: &mut VecDeque<(i64, i64)>, input: &[(i64, i64)]) {
     for (key, _) in input {
         let idx = deq.iter().position(|(orig, _)| *orig == *key).unwrap() as i64;
         let value = deq.remove(idx as usize).unwrap();
@@ -11,13 +11,13 @@ fn run_once(deq: &mut VecDeque<(i64, i128)>, input: &[(i64, i128)]) {
     }
 }
 
-fn mix(vals: &[i128], times: usize) -> i128 {
-    let input: Vec<(i64, i128)> = vals
+fn mix(vals: &[i64], times: usize) -> i64 {
+    let input: Vec<(i64, i64)> = vals
         .iter()
         .enumerate()
         .map(|(i, v)| (i as i64, *v))
         .collect();
-    let mut deq: VecDeque<(i64, i128)> = input.iter().copied().collect();
+    let mut deq: VecDeque<(i64, i64)> = input.iter().copied().collect();
     (0..times).for_each(|_| run_once(&mut deq, &input));
     let zero = deq.iter().position(|(_, v)| *v == 0).unwrap();
     deq.iter()
@@ -26,11 +26,11 @@ fn mix(vals: &[i128], times: usize) -> i128 {
         .take_while(|(i, _)| *i <= zero + 3000)
         .filter(|(i, _)| *i == zero + 1000 || *i == zero + 2000 || *i == zero + 3000)
         .map(|(_, (_, v))| *v)
-        .sum::<i128>()
+        .sum::<i64>()
 }
 
 pub fn solve(input: &str, verify_expected: bool, output: bool) -> Result<Duration> {
-    let mut input: Vec<i128> = input.lines().map(|s| s.parse().unwrap()).collect();
+    let mut input: Vec<i64> = input.lines().map(|s| s.parse().unwrap()).collect();
 
     let s = Instant::now();
